* Update directory * Update Apache 2 license
