+++ examples/cloud-datafusion-functions-plugins/img/1-create_pipeline_checkpoint_collection.png
