+++ examples/cloud-datafusion-functions-plugins/img/11-truncate_table_action_ui_pipeline_parameters.png
