+++ examples/cloud-datafusion-functions-plugins/img/12-database_source_ui_pipeline_parameters.png
