+++ examples/cloud-datafusion-functions-plugins/img/13-bigquery_sink_ui_pipeline_parameters.png
