+++ examples/cloud-datafusion-functions-plugins/img/14-merge_last_update_ts_action.png
