+++ examples/cloud-datafusion-functions-plugins/img/15-checkpoint_update_action_ui_pipeline_parameters.png
