+++ examples/cloud-datafusion-functions-plugins/img/2-create_check_point_collection.png
