+++ examples/cloud-datafusion-functions-plugins/img/3-runtime_arguments.png
