+++ examples/cloud-datafusion-functions-plugins/img/4-checkpoint_read_action_plugin_ui.png
