+++ examples/cloud-datafusion-functions-plugins/img/5-checkpoint_update_action_plugin_ui.png
