+++ examples/cloud-datafusion-functions-plugins/img/6-copy_table_action_ui.png
