+++ examples/cloud-datafusion-functions-plugins/img/7-drop_table_action_ui.png
