+++ examples/cloud-datafusion-functions-plugins/img/8-truncate_table_action_ui.png
