+++ examples/cloud-datafusion-functions-plugins/img/9-set_runtime_arugments_latestWatermarkValue.png
