+++ examples/cloud-datafusion-functions-plugins/img/9-successful_run_incremental_v2_pipeline.png
