+++ examples/cloud-datafusion-functions-plugins/src/main/java/com/google/functions/InsertOverwriteFunction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Google Inc.
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
