+++ examples/cloud-datafusion-functions-plugins/widgets/CheckPointReadAction-action.json
