+++ examples/cloud-datafusion-functions-plugins/widgets/CheckPointUpdateAction-action.json
