+++ examples/cloud-datafusion-functions-plugins/widgets/CopyTableAction-action.json
