+++ examples/cloud-datafusion-functions-plugins/widgets/DropTableAction-action.json
