+++ examples/cloud-datafusion-functions-plugins/widgets/TruncateTableAction-action.json
